ocr_inspector/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pymupdf
pytesseract
